requirements.txt
tensorflow
numpy
Pillow
tqdm